Update tests for new Celery-based API
