+++ tests/e2e/test_api.py
@@ -11,63 +11,12 @@ from document.entrypoints.app import app
 from fastapi.testclient import TestClient
 
 from document.domain import model
+from tests.shared.utils import (
+    check_finished_document_with_verses_success,
+    check_finished_document_without_verses_success,
+)
 
 
-def check_finished_document_with_verses_success(response: requests.Response) -> None:
-    """
-    Helper to keep tests DRY.
-
-    Check that the finished_document_path exists and also check that
-    the HTML file associated with it exists and includes verses_html.
-    """
-    assert response.ok
-    content = response.json()
-    assert "finished_document_request_key" in content
-    assert "message" in content
-    html_filepath = os.path.join(
-        settings.output_dir(),
-        "{}.html".format(content["finished_document_request_key"]),
-    )
-    assert response.json() == {
-        "finished_document_request_key": content["finished_document_request_key"],
-        "message": settings.SUCCESS_MESSAGE,
-    }
-    with open(html_filepath, "r") as fin:
-        html = fin.read()
-        parser = bs4.BeautifulSoup(html, "html.parser")
-        body = parser.find_all("body")
-        assert body
-        verses_html = parser.find_all("span", attrs={"class": "v-num"})
-        assert verses_html
-    assert response.ok
-
-
-def check_finished_document_without_verses_success(response: requests.Response) -> None:
-    """
-    Helper to keep tests DRY.
-
-    Check that the finished_document_path exists and also check that
-    the HTML file associated with it exists and includes body but not
-    verses_html.
-    """
-    assert response.ok
-    content = response.json()
-    assert "finished_document_request_key" in content
-    assert "message" in content
-    html_filepath = os.path.join(
-        settings.output_dir(),
-        "{}.html".format(content["finished_document_request_key"]),
-    )
-    with open(html_filepath, "r") as fin:
-        html = fin.read()
-        parser = bs4.BeautifulSoup(html, "html.parser")
-        body = parser.find_all("body")
-        assert body
-        verses_html = parser.find_all("span", attrs={"class": "v-num"})
-        # reg is malformed and udb does not exist, thus there is
-        # no html generated
-        assert not verses_html
-    assert response.ok
 
 
 ##########################################################################
@@ -88,9 +37,9 @@ def test_en_ulb_wa_col_en_tn_wa_col_language_book_order_with_no_email_1c() -> No
                 "assembly_strategy_kind": model.AssemblyStrategyEnum.LANGUAGE_BOOK_ORDER,
                 "assembly_layout_kind": model.AssemblyLayoutEnum.ONE_COLUMN,
                 "layout_for_print": False,
-                "generate_pdf": True,
+                "generate_pdf": False,
                 "generate_epub": False,
-                "generate_docx": False,
+                "generate_docx": True,
                 "resource_requests": [
                     {
                         "lang_code": "en",
@@ -105,7 +54,7 @@ def test_en_ulb_wa_col_en_tn_wa_col_language_book_order_with_no_email_1c() -> No
                 ],
             },
         )
-        check_finished_document_with_verses_success(response)
+        check_finished_document_with_verses_success(response, suffix="docx")
 
 
 def test_en_ulb_wa_col_en_tn_wa_col_language_book_order_with_no_email_1c_c() -> None:
